Move labels into lib/utils
